[#4] feat. image 호출 파일 분리
