+++ src/utils/UI/equip.ts
@@ -1,0 +1,19 @@
+import equip1 from "../../image/item/etc/04000001.info.icon.png";
+import equip2 from "../../image/item/etc/04000002.info.icon.png";
+import equip3 from "../../image/item/etc/04000003.info.icon.png";
+import equip4 from "../../image/item/etc/04000004.info.icon.png";
+import equip5 from "../../image/item/etc/04000005.info.icon.png";
+import equip6 from "../../image/item/etc/04000006.info.icon.png";
+import equip7 from "../../image/item/etc/04000007.info.icon.png";
+import equip8 from "../../image/item/etc/04000008.info.icon.png";
+
+export const EquipImages = [
+  equip1,
+  equip2,
+  equip3,
+  equip4,
+  equip5,
+  equip6,
+  equip7,
+  equip8,
+];
