+++ src/utils/UI/inventory.ts
@@ -1,0 +1,30 @@
+import inventory from "../../image/UI/inventory/Item.backgrnd.png";
+import inventory3 from "../../image/UI/inventory/Item.backgrnd3.png";
+import inventory2 from "../../image/UI/inventory/Item.backgrnd2.png";
+import invenEquipDisabled from "../../image/UI/inventory/Item.Tab.disabled.0.png";
+import invenUseDisabled from "../../image/UI/inventory/Item.Tab.disabled.1.png";
+import invenEtcDisabled from "../../image/UI/inventory/Item.Tab.disabled.2.png";
+import invenSetupDisabled from "../../image/UI/inventory/Item.Tab.disabled.3.png";
+import invenEquipEnabled from "../../image/UI/inventory/Item.Tab.enabled.0.png";
+import invenUseEnabled from "../../image/UI/inventory/Item.Tab.enabled.1.png";
+import invenEtcEnabled from "../../image/UI/inventory/Item.Tab.enabled.2.png";
+import invenSetupEnabled from "../../image/UI/inventory/Item.Tab.enabled.3.png";
+
+import wood_weapon from "../../image/item/etc/04000003.info.icon.png";
+import burble_weapon from "../../image/item/etc/04000011.info.icon.png";
+
+export const images = {
+  inventory,
+  inventory2,
+  inventory3,
+  invenEquipDisabled,
+  invenUseDisabled,
+  invenEtcDisabled,
+  invenSetupDisabled,
+  invenEquipEnabled,
+  invenUseEnabled,
+  invenEtcEnabled,
+  invenSetupEnabled,
+  wood_weapon,
+  burble_weapon,
+};
