+++ src/utils/UI/tooltip.ts
@@ -1,0 +1,39 @@
+import tooltip_flame_top from "../../image/UI/ItemToolTip/Frame/Item.Frame.top.png";
+import tooltip_flame_bottum from "../../image/UI/ItemToolTip/Frame/Item.Frame.bottom.png";
+import tooltip_flame_dotline from "../../image/UI/ItemToolTip/Frame/Item.Frame.dotline.png";
+import tooltip_flame_line from "../../image/UI/ItemToolTip/Frame/Item.Frame.line.png";
+import tooltip_flame_cover from "../../image/UI/ItemToolTip/Frame/Item.Frame.cover.png";
+
+import tooltip_Item_Icon_rare from "../../image/UI/ItemToolTip/ItemIcon/Item.ItemIcon.1.png";
+import tooltip_Item_Icon_epic from "../../image/UI/ItemToolTip/ItemIcon/Item.ItemIcon.2.png";
+import tooltip_Item_Icon_unique from "../../image/UI/ItemToolTip/ItemIcon/Item.ItemIcon.3.png";
+import tooltip_Item_Icon_legendary from "../../image/UI/ItemToolTip/ItemIcon/Item.ItemIcon.4.png";
+import tooltip_Item_Icon_god from "../../image/UI/ItemToolTip/ItemIcon/Item.ItemIcon.0.png";
+import tooltip_Item_Icon_base from "../../image/UI/ItemToolTip/ItemIcon/Item.ItemIcon.base.png";
+import tooltip_Item_Icon_cover from "../../image/UI/ItemToolTip/ItemIcon/Item.ItemIcon.cover.png";
+import tooltip_Item_Icon_shade from "../../image/UI/ItemToolTip/ItemIcon/Item.ItemIcon.shade.png";
+
+import tooltip_Item_Star from "../../image/UI/ItemToolTip/Equip/Star/Item.Equip.Star.Star.png";
+import tooltip_Item_Star_none from "../../image/UI/ItemToolTip/Equip/Star/Item.Equip.Star.Star0.png";
+import tooltip_Item_Star_blue from "../../image/UI/ItemToolTip/Equip/Star/Item.Equip.Star.Star1.png";
+
+export const tooltip = {
+  tooltip_flame_top,
+  tooltip_flame_bottum,
+  tooltip_flame_dotline,
+  tooltip_flame_line,
+  tooltip_flame_cover,
+
+  tooltip_Item_Icon_rare,
+  tooltip_Item_Icon_epic,
+  tooltip_Item_Icon_unique,
+  tooltip_Item_Icon_legendary,
+  tooltip_Item_Icon_god,
+  tooltip_Item_Icon_base,
+  tooltip_Item_Icon_cover,
+  tooltip_Item_Icon_shade,
+
+  tooltip_Item_Star,
+  tooltip_Item_Star_none,
+  tooltip_Item_Star_blue,
+};
