+++ src/utils/images.ts
@@ -1,50 +1,9 @@
-import inventory from "../image/UI/Item.backgrnd.png";
-import inventory3 from "../image/UI/Item.backgrnd3.png";
-import inventory2 from "../image/UI/Item.backgrnd2.png";
-import invenEquipDisabled from "../image/UI/Item.Tab.disabled.0.png";
-import invenUseDisabled from "../image/UI/Item.Tab.disabled.1.png";
-import invenEtcDisabled from "../image/UI/Item.Tab.disabled.2.png";
-import invenSetupDisabled from "../image/UI/Item.Tab.disabled.3.png";
-import invenEquipEnabled from "../image/UI/Item.Tab.enabled.0.png";
-import invenUseEnabled from "../image/UI/Item.Tab.enabled.1.png";
-import invenEtcEnabled from "../image/UI/Item.Tab.enabled.2.png";
-import invenSetupEnabled from "../image/UI/Item.Tab.enabled.3.png";
+import { EquipImages } from "./UI/equip";
+import { images } from "./UI/inventory";
+import { tooltip } from "./UI/tooltip";
 
-import wood_weapon from "../image/item/etc/04000003.info.icon.png";
-import burble_weapon from "../image/item/etc/04000011.info.icon.png";
-
-import equip1 from "../image/item/etc/04000001.info.icon.png";
-import equip2 from "../image/item/etc/04000002.info.icon.png";
-import equip3 from "../image/item/etc/04000003.info.icon.png";
-import equip4 from "../image/item/etc/04000004.info.icon.png";
-import equip5 from "../image/item/etc/04000005.info.icon.png";
-import equip6 from "../image/item/etc/04000006.info.icon.png";
-import equip7 from "../image/item/etc/04000007.info.icon.png";
-import equip8 from "../image/item/etc/04000008.info.icon.png";
-
-export const images = {
-  inventory,
-  inventory2,
-  inventory3,
-  invenEquipDisabled,
-  invenUseDisabled,
-  invenEtcDisabled,
-  invenSetupDisabled,
-  invenEquipEnabled,
-  invenUseEnabled,
-  invenEtcEnabled,
-  invenSetupEnabled,
-  wood_weapon,
-  burble_weapon,
+export default {
+  EquipImages,
+  images,
+  tooltip,
 };
-
-export const EquipImages = [
-  equip1,
-  equip2,
-  equip3,
-  equip4,
-  equip5,
-  equip6,
-  equip7,
-  equip8,
-];
